requirements.txt
# Core Framework
fastapi>=0.104.1,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-socketio>=5.10.0,<6.0.0
python-engineio>=4.8.0,<5.0.0

# Database
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.12.1,<2.0.0

# Redis & Task Queue
redis>=5.0.1,<6.0.0
celery>=5.3.4,<6.0.0

# Web Scraping & Automation
selenium>=4.15.2,<5.0.0
playwright>=1.40.0,<2.0.0
beautifulsoup4>=4.12.2,<5.0.0
requests>=2.31.0,<3.0.0
lxml>=4.9.3,<5.0.0

# Domain & Network Analysis
dnspython>=2.4.2,<3.0.0
python-whois>=0.8.0,<1.0.0
whois>=0.9.27,<1.0.0

# Image Processing & Similarity
opencv-python>=4.8.1.78,<5.0.0
Pillow>=10.1.0,<11.0.0
scikit-image>=0.22.0,<1.0.0
imagehash>=4.3.1,<5.0.0

# PDF Generation
reportlab>=4.0.7,<5.0.0
pypdf>=3.17.1,<4.0.0

# Data Processing
pandas>=2.1.3,<3.0.0
numpy>=1.26.2,<2.0.0

# Machine Learning
scikit-learn>=1.3.2,<2.0.0
xgboost>=2.0.2,<3.0.0
lightgbm>=4.1.0,<5.0.0
joblib>=1.3.2,<2.0.0

# Natural Language Processing
nltk>=3.8.1,<4.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Security & Validation
validators>=0.22.0,<1.0.0
tldextract>=5.1.0,<6.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-multipart>=0.0.6,<1.0.0

# Process Management
psutil>=5.9.0,<6.0.0

# Logging
loguru>=0.7.2,<1.0.0

# Date handling
python-dateutil>=2.8.2,<3.0.0

# Additional dependencies for stability
urllib3>=1.26.0,<2.0.0
certifi>=2023.0.0
charset-normalizer>=3.0.0

